+++ NEWS.md
@@ -18,6 +18,7 @@ Bug fixes:
 + Fix issue that would make Phan infer that a redundant/impossible condition outside a loop was in a loop.
 + Avoid false positives analyzing expressions within `assert()`.
 + Fix method signatures for php 7.4's `WeakReference`.
++ Fix false positives analyzing uses of `__call` and `__callStatic` (#702)
 
 Jun 14 2019, Phan 2.2.0
 -----------------------